requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.3.0
transformers==4.41.1
torch==2.2.2+cpu
tqdm==4.66.4
pandas==2.2.2
scikit-learn==1.5.0
python-dotenv==1.0.1
